+++ src/file/file.cpp
@@ -5,6 +5,7 @@
 #endif
 
 #include <invader/file/file.hpp>
+#include <invader/error.hpp>
 #include <invader/printf.hpp>
 
 #include <cstdio>
@@ -103,7 +104,29 @@ namespace Invader::File {
     }
 
     std::filesystem::path tag_path_to_file_path(const std::string &tag_path, const std::filesystem::path &tags) {
-        return tags / halo_path_to_preferred_path(tag_path);
+        std::filesystem::path p = tag_path;
+        auto path = tags / halo_path_to_preferred_path(tag_path);
+        
+        // Check if an absolute path or if it uses . or .. path components (which can potentially cause some bad traversal)
+        bool malicious_maybe = false;
+        if(p.is_absolute()) {
+            malicious_maybe = true;
+        }
+        else {
+            for(auto &c : path) {
+                auto str = c.string();
+                if(str == "." || str == "..") {
+                    malicious_maybe = true;
+                }
+            }
+        }
+        
+        // Path is probably bad
+        if(malicious_maybe) {
+            throw InvalidTagPathException();
+        }
+        
+        return path;
     }
 
     std::optional<std::string> file_path_to_tag_path(const std::filesystem::path &file_path, const std::filesystem::path &tags) {